use the not equals judge for seq no.
fix this case:
1. the other extension(say, custom script) is called.
2. the wala save the settings file for the vmbackup extension.
3. the extension would be called again.

so we need to use the not equals when judge the seq no.
also for the restore case. the seqNo should not be equals to the last
one success.
one fail case here is the 0. this case is mitigated by the machine
identity, and then seq no set to -1.