Fix for Internal DNS Check
I needed the ability to disable the internal DNS check, and while the public_setting is exposed, there is a bug where it was not being handled correctly and didn't work.  Setting a var to a string value makes it a 'true' boolean value when used for a boolean operation regardless of what the string value actually is.

This change uses the strtobool function to convert the string value that is optionally defined in public_settings ('enableInternalDNSCheck') into a true boolean value for the 'enable_idns_check' variable, so that the idns check is executes as expected.  Omitting 'enableInternalDNSCheck' from the public_settings results in a default of 'true' (as it was) while allowing a string value of 'y, yes, t, true, on or 1' for boolean true and a string value of 'n, no, f, false, off and 0' as a boolean false.