Simple Makefile to build extensions.
Use the version and type (name) from an extension's manifest.xml to
create an extension's zip archive.

The Makefile only supports only the extensions that I care about for the
time being.  There is considerably more work necessary to clean the rest
up, and I will pay that cost eventually, but not today.