Further consolidate Utils.
Consolidate more of the various copies of Utils to the global copy.  The
major extensions have now been moved over.

When HandlerUtil is constructed attempt to determine the extension's long
name, short name, and verison by parsing the manifest or by examing the
current working directory.

Remove any calls to determine the name by checking HandlerManifest.json.
Name is not a recognized schema element.

Remove any calls to determine the version by checking
HandlerManifest.json.  Version is a recognized element, but it determines
the schema version, not the extension version.  It is improper to do so.