Optimize system api calls for size calculation (#1677)
With the new version (1.0.9197) release of Linux VmSnapshot, which has the changes to support excluded disks during billing calculation, the newly introduced system commands (lsblk, pv, lsscsi) were executed even when the no disks are excluded. This happens because newly added system commands are run during initialization of SizeCalculation.
To optimize the behavior, made changes in SizeCalculation.py (at line 181) so that system commands are run only when one or more disks are excluded.